pull out RTC renderer
